+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * `cases` kwarg to `plot_2d_cov_ellipse`, `plot_integration_error`, and `plot_integration_convergence` to allow specifying which cases to use for the plot
 ### Changed    
 * Switch varstatbootstrap method from 'basic' to 'BCa'
+* `get_cases` function moved from `mc_plot` to `helper_functions`
 ### Removed    
 
 ## [0.14.1] - 2025-01-31